Extract partition::PartitionMap into own module.
